requirements.txt
streamlit
langchain
langchain-community
openai
pypdf
tiktoken
python-dotenv
langchain-text-splitters
